requirements.txt
pdf2image
pytesseract
pillow
transformers
torch
torchvision
torchaudio
git+https://github.com/facebookresearch/detectron2.git